added AutoIncrement to Model with an int as its primary key

+++ Common/ServiceStack.Common/ServiceStack.OrmLite.Tests/Models/ModelWithIdOnly.cs
@@ -6,12 +6,13 @@ namespace ServiceStack.OrmLite.Tests.Models
 		{
 		}
 
-		public ModelWithIdOnly(int id)
+		public ModelWithIdOnly(long id)
 		{
 			Id = id;
 		}
 
-		public int Id { get; set; }
+		// must be long as you cannot have a table with only an autoincrement field
+		public long Id { get; set; }
 
 	}
 }
